+++ console/cypress/integration/actions/spec.ts
@@ -95,9 +95,9 @@ const clickOnCreateAction = () => {
 };
 
 export const routeToGraphiql = () => {
-  cy.visit('/api-explorer');
+  cy.visit('/api/api-explorer');
   cy.wait(7000);
-  cy.url().should('eq', `${baseUrl}/api-explorer`);
+  cy.url().should('eq', `${baseUrl}/api/api-explorer`);
 };
 
 export const createMutationAction = () => {