+++ console/cypress/integration/data/computed-fields/spec.ts
@@ -124,9 +124,9 @@ export const openModifySection = () => {
 };
 
 export const routeToGraphiql = () => {
-  cy.visit('/api-explorer');
+  cy.visit('/api/api-explorer');
   cy.wait(7000);
-  cy.url().should('eq', `${baseUrl}/api-explorer`);
+  cy.url().should('eq', `${baseUrl}/api/api-explorer`);
 };
 
 export const verifyComputedFieldsResult = () => {